+++ src/components/Profile/Skills.js
@@ -2,14 +2,31 @@ import React from 'react';
 import {connect} from 'react-redux';
 import {deleteFromDatabase} from '../../redux/actions/actions';
 import removeIcon from '../../assets/delete-icon-white.png';
+import { confirmAlert } from 'react-confirm-alert';
+
+import 'react-confirm-alert/src/react-confirm-alert.css';
+
+
 
 
 const Skills = ({Data: {skill, level, id}, enableEdit, user, deleteFromDatabase}) => {
   const removeSkill = () => {
-    if (user) {
-      const path = `skills/${user.uid}`
-      deleteFromDatabase(path, id)
-    }
+
+
+    confirmAlert({
+        title: 'Confirm to remove',
+        message: 'Are you sure to do this.',
+        buttons: [{
+            label: 'Yes',
+            onClick: () => {
+              if (user) {
+                const path = `skills/${user.uid}`
+                deleteFromDatabase(path, id)
+              }
+            }
+          },
+          { label: 'No'}
+        ]})
   }
   return (
     <div className="section__body__skills">
